Vue3 Products++++++

+++ app/Modules/Accounting/Controllers/ArrivalController.php
@@ -173,7 +173,7 @@ class ArrivalController extends Controller
     public function add_product(Request $request, ArrivalDocument $arrival): RedirectResponse
     {
         try {
-            $this->service->addProduct($arrival, $request->integer('product_id'), $request->integer('quantity'));
+            $this->service->addProduct($arrival, $request->integer('product_id'), $request->float('quantity'));
             return redirect()->back()->with('success', 'Товар добавлен');
         } catch (\DomainException $e) {
             return redirect()->back()->with('error', $e->getMessage());
