+++ app/Modules/Accounting/Entity/ArrivalDocument.php
@@ -76,15 +76,15 @@ class ArrivalDocument extends AccountingDocument
         $amount = ArrivalProduct::selectRaw('SUM(quantity * cost_currency) AS total')
             ->where('arrival_id', $this->id)
             ->first();
-        return (int)$amount->total ?? 0.0;
+        return (float)$amount->total ?? 0.0;
     }
 
-    public function getQuantity(): int
+    public function getQuantity(): float
     {
         $quantity = ArrivalProduct::selectRaw('SUM(quantity * 1) AS total')
             ->where('arrival_id', $this->id)
             ->first();
-        return (int)($quantity->total ?? 0);
+        return (float)($quantity->total ?? 0);
     }
 
     public function getManager(): string
