+++ app/Modules/Accounting/Entity/DepartureProduct.php
@@ -21,7 +21,7 @@ class DepartureProduct extends AccountingProduct
         'cost',
     ];
 
-    public static function new(int $product_id, int $quantity, float $cost): self
+    public static function new(int $product_id, float $quantity, float $cost): self
     {
         $product = self::baseNew($product_id, $quantity);
         $product->cost = $cost;
